requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
attrs==24.2.0
backoff==2.2.1
bcrypt==4.3.0
beautifulsoup4==4.13.5
build==1.3.0
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
chromadb==1.0.20
click==8.2.1
colorama==0.4.6
coloredlogs==15.0.1
dataclasses-json==0.6.7
decorator==5.1.1
distro==1.9.0
Django==4.2.20
django-cors-headers==4.7.0
djangorestframework==3.16.0
# Editable install with no version control (documentation==0.1.0)
# -e d:\projects\knowledge extractor\mcp-server-example
durationpy==0.10
filelock==3.16.1
flatbuffers==25.2.10
frozenlist==1.7.0
fsspec==2024.12.0
google-ai-generativelanguage==0.6.15
google-api-core==2.24.0
google-api-python-client==2.159.0
google-auth==2.37.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.3
google-cloud-discoveryengine==0.13.11
google-cloud-documentai==3.5.0
google-cloud-storage==3.3.1
google-crc32c==1.7.1
google-genai==1.20.0
google-generativeai==0.8.5
google-resumable-media==2.7.2
googleapis-common-protos==1.66.0
greenlet==3.2.4
grpcio==1.73.1
grpcio-status==1.71.2
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.1
huggingface-hub==0.27.1
humanfriendly==10.0
idna==3.10
imageio==2.36.1
imageio-ffmpeg==0.5.1
importlib_metadata==8.7.0
importlib_resources==6.5.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.4.1
kubernetes==33.1.0
langchain==0.3.27
langchain-community==0.3.29
langchain-core==0.3.75
langchain-text-splitters==0.3.10
langsmith==0.4.21
markdown-it-py==4.0.0
marshmallow==3.26.1
mcp==1.13.1
mdurl==0.1.2
mmh3==5.2.0
moviepy==2.1.2
mpmath==1.3.0
multidict==6.6.4
mypy_extensions==1.1.0
numpy==2.2.1
oauthlib==3.3.1
onnxruntime==1.22.1
opentelemetry-api==1.36.0
opentelemetry-exporter-otlp-proto-common==1.36.0
opentelemetry-exporter-otlp-proto-grpc==1.36.0
opentelemetry-proto==1.36.0
opentelemetry-sdk==1.36.0
opentelemetry-semantic-conventions==0.57b0
orjson==3.11.3
outcome==1.3.0.post0
overrides==7.7.0
packaging==24.2
pathspec==0.11.2
pillow==10.4.0
posthog==5.4.0
proglog==0.1.10
propcache==0.3.2
proto-plus==1.25.0
protobuf==5.29.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pybase64==1.4.2
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
Pygments==2.19.2
pyparsing==3.2.1
PyPika==0.48.9
pyproject_hooks==1.2.0
pyreadline3==3.5.4
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
pywin32==311
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.5
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==14.1.0
rpds-py==0.27.1
rsa==4.9
safetensors==0.5.2
selenium==4.27.1
setuptools==75.8.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.8
SQLAlchemy==2.0.43
sqlparse==0.5.3
sse-starlette==3.0.2
starlette==0.47.3
sympy==1.14.0
tenacity==9.1.2
tokenizers==0.21.0
tqdm==4.67.1
transformers==4.48.0
trio==0.27.0
trio-websocket==0.11.1
typer==0.17.3
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.12.2
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.2.3
uvicorn==0.35.0
watchfiles==1.1.0
webdriver-manager==4.0.2
websocket-client==1.8.0
websockets==15.0.1
whitenoise==6.9.0
wsproto==1.2.0
yarl==1.20.1
zipp==3.23.0
zstandard==0.24.0
